Correctly add time-delayed restore
PyGObject lacking `idle_add_once` and `timeout_add_once`, the
appropriate pattern is to use `timeout_add` and return `False` from the
called function which signals the main loop to remove the timeout
instead of running it again when the timeout expires.

Note that the units for `timeout_add`'s interval is milliseconds where
both `time.sleep` and `threading.Timer` are both in seconds.